added fetchNFT with price as well
having issue with fetching price it is not loading as soon as page is reloaded, some working is need to be done on getPrice function that called from the smart contract
